+++ src/mesido/esdl/asset_to_component_base.py
@@ -204,7 +204,7 @@ class _AssetToComponentBase:
         "CheckValve": "check_valve",
         "Compressor": "compressor",
         "CoolingDemand": "cold_demand",
-        "ElectricBoiler": "elec_boiler",
+        "ElectricBoiler": "heat_source_elec",
         "ElectricityCable": "electricity_cable",
         "ElectricityDemand": "electricity_demand",
         "ElectricityProducer": "electricity_source",
@@ -251,7 +251,8 @@ class _AssetToComponentBase:
             "fixedMaintenanceCosts": "optional",
             "fixedOperationalCosts": "optional",
         },
-        "heat_source": {  # Includes GeothermalSource, ResidualHeatSource, HeatProducer
+        "heat_source": {  # Includes GeothermalSource, ResidualHeatSource, HeatProducer,
+            # ElectricBoiler
             "investmentCosts": "required",
             "installationCosts": "required",
             "variableOperationalCosts": "required",
@@ -302,6 +303,7 @@ class _AssetToComponentBase:
         "ATES": "ates",
         "CoolingDemand": "heat_demand",
         "Electrolyzer": "electrolyzer",
+        "ElectricBoiler": "heat_source",
         "GasDemand": "gas_demand",
         "GasStorage": "gas_tank_storage",
         "GeothermalSource": "heat_source",
@@ -1020,7 +1022,9 @@ class _AssetToComponentBase:
                     if q_nominal is not None:
                         self._set_q_nominal(asset, q_nominal)
                         return q_nominal
-        elif len(asset.in_ports) == 2 and len(asset.out_ports) == 1:  # for gas_boiler or e_boiler
+        elif (
+            len(asset.in_ports) == 2 and len(asset.out_ports) == 1
+        ):  # for gas_boiler or elec_heat_source_elec
             q_nominals = {}
             try:
                 for port in asset.in_ports: